feat(db): migrate from Planetscale Postgres to Cloudflare D1
- schema.ts ported to sqlite-core (autoincrement PKs, text json, epoch
  int timestamps, 0/1 booleans); fresh 0000_d1_baseline migration
- drizzle.config dialect: sqlite
- alchemy.run.ts: bind D1Database (auto read replication, apac primary),
  drop Hyperdrive + Singapore placement
- DB client layer -> drizzle-orm/d1; drop postgres-js + DIRECT_DB_URL
- rewrite interactive transactions to sequential ops (D1 has none):
  checkout, payment confirmation, purchase CRUD, ai-purchase import
- ilike -> like; drop Postgres ::int casts; NOW() -> unixepoch()

+++ packages/api/src/queries/brands.ts
@@ -5,7 +5,7 @@ import { BrandsTable, ProductsTable } from "~/db/schema";
 export const brandQueries = {
 	admin: {
 		async getAllBrands() {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			return db()
 				.select({
@@ -88,7 +88,7 @@ export const brandQueries = {
 
 	store: {
 		async getAllBrands() {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			return db()
 				.select({
@@ -120,7 +120,7 @@ export const brandQueries = {
 		},
 
 		async getBrandById(id: number) {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			const result = await db()
 				.select({
@@ -162,7 +162,7 @@ export const brandQueries = {
 		},
 
 		async getBrandBySlug(slug: string) {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			const result = await db()
 				.select({
@@ -204,7 +204,7 @@ export const brandQueries = {
 		},
 
 		async getAllBrandsWithStock() {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			return db()
 				.select({
